+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Dedicated Goodreads and custom CSV import flows in the Stacks page with better
+  instructions for each.
+
 ## [1.2.1] - 2026-06-17
 
 ### Changed